% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{StackR_Miletus}
\alias{StackR_Miletus}
\title{Stacking Ring from Miletus}
\format{
A data frame with 67 rows and 15 variables
}
\source{
https://doi.org/10.34780/aa.v0i1.1014
}
\usage{
data(StackR_Miletus)
}
\description{
This data contains all individual rings listed in fig. 16 from Finds from
"Miletus XXXII: Clay Rings from the Sanctuary of Dionysos in Miletus,
Archäologischer Anzeiger 2020/1" (https://doi.org/10.34780/aa.v0i1.1014).
}
\details{
\itemize{
  \item Inv: Inventory Number in Miletus. If NA not inventoried.
  \item clay: Color of the clay (Names as used in the Munsell Soil Color Chart)
  \item applications: Type of applications
  \item color_appl: Color of applications (Names as used in the Munsell Soil Color Chart)
  \item markings: Markings on the surfaces
  \item height: Height of the ring body in cm
  \item min_DM: Minimum (inner) diameter in cm
  \item max_DM: Maximum (outer) diameter in cm
  \item width: Surface width of the clay ring in cm (distance between min_DM and max_DM depending on angle)
  \item thickness: Thickness of the clay in cm (taken at the most representative spot)
  \item shape: Overall shape of the ring body in profile
  \item fragments: Number of fragments assigned to this ring
}
}
\keyword{datasets}
